refactor: extract shared OpenCode-format logic into opencode_common module
Kilo CLI is a fork of OpenCode and both use the identical on-disk format
(one JSON file per message, per-session directories). The two analyzer
implementations were ~93% duplicated (~1,464 lines of near-identical code).

Extract all shared types, helpers, and the Analyzer trait implementation
into a new opencode_common module parameterised by an OpenCodeFormatConfig
(display name, Application variant, hash prefix, storage sub-directory).
Both OpenCodeAnalyzer and KiloCliAnalyzer are now thin newtype wrappers
that delegate to OpenCodeFormatAnalyzer with their respective configs.

Before: opencode.rs (681 lines) + kilo_cli.rs (867 lines) = 1548 lines
After:  opencode_common.rs (949) + opencode.rs (66) + kilo_cli.rs (66) = 1081 lines

All 215 tests pass, clippy clean, fmt clean, docs build.

Addresses review feedback from Sewer56 on PR #118.
